requirements.txt
Django==1.10.5
django-crispy-forms==1.6.1
django-filter==1.0.1
djangorestframework==3.5.4
Markdown==2.6.8
